backend_jobApp/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.6
django-braces==1.14.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-hosts==4.0
django-js-asset==1.2.2
django-oauth-toolkit==1.5.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.2
phonenumberslite==8.12.29
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.12.1
